update view into inbox.view
