+++ socialdisto/urls.py
@@ -35,5 +35,5 @@ from django.urls import include, path
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('accounts.urls')),
-    path('inbox/', include('inbox.urls'))
+    path('inbox/', include('inbox.urls')),
 ]
